add to project

+++ app/Http/Controllers/admin/AdminController.php
@@ -15,35 +15,59 @@ use phpDocumentor\Reflection\Types\Integer;
 
 class AdminController extends Controller
 {
+
+
     public function index()
     {
+        if (User::where('id',Auth::id())->first('role')['role']!='admin'){
+            redirect('/');
+        }
         return view('admin/panel');
     }
 
     public function user()
     {
+        if (User::where('id',Auth::id())->first('role')['role']!='admin'){
+            redirect('/');
+        }
         return view('admin/user', ['user' => User::all()]);
     }
-    public function userDelete($id){
+
+    public function userDelete($id)
+    {
+        if (User::where('id',Auth::id())->first('role')['role']!='admin'){
+            redirect('/');
+        }
         User::destroy($id);
         return redirect()->back();
     }
-    public function userAdmin($id){
-        $user=User::find($id);
+
+    public function userAdmin($id)
+    {
+        if (User::where('id',Auth::id())->first('role')['role']!='admin'){
+            redirect('/');
+        }
+        $user = User::find($id);
         $user->update([
-            'role'=>'admin'
+            'role' => 'admin'
         ]);
         return redirect()->back();
     }
 
     public function Exam()
     {
+        if (User::where('id',Auth::id())->first('role')['role']!='admin'){
+            redirect('/');
+        }
 
         return view('admin/examPanel', ['Exam' => Test::all()]);
     }
 
     public function testCreate()
     {
+        if (User::where('id',Auth::id())->first('role')['role']!='admin'){
+            redirect('/');
+        }
         return view('admin/creatTest');
 
     }
@@ -73,6 +97,9 @@ class AdminController extends Controller
 
     public function examCreate($id)
     {
+        if (User::where('id',Auth::id())->first('role')['role']!='admin'){
+            redirect('/');
+        }
         $r = Test::where('id', $id)->first()['exam'];
         return \view('admin/examCreate', ['id' => $id, 'r' => $r]);
     }
@@ -80,7 +107,6 @@ class AdminController extends Controller
     public function examStore($id)
     {
         $r = Test::where('id', $id)->first()['exam'];
-//        return $r;
         if (!Exam::where('test_id', $id)->get() || count(Exam::where('test_id', $id)->get()) == 0) {
             for ($x = 1; $x <= $r; $x++) {
                 $validate = \request()->validate([
@@ -88,9 +114,23 @@ class AdminController extends Controller
                     "answer_{$x}" => 'required',
                 ]);
                 redirect()->back()->withErrors($validate);
+                $test = [];
+                if (strlen(\request("test_{$x}_1")) != 0 || \request("answer_{$x}") =="test_{$x}_1") {
+                    array_push($test, \request("test_{$x}_1"));
+                };
+                if (strlen(\request("test_{$x}_2")) != 0 || \request("answer_{$x}") =="test_{$x}_2") {
+                    array_push($test, \request("test_{$x}_2"));
+                };
+                if (strlen(\request("test_{$x}_3")) != 0 || \request("answer_{$x}") =="test_{$x}_3") {
+                    array_push($test, \request("test_{$x}_3"));
+                };
+                if (strlen(\request("test_{$x}_4")) != 0 || \request("answer_{$x}") =="test_{$x}_4") {
+                    array_push($test, \request("test_{$x}_4"));
+                };
                 Exam::create([
-                    'qu' => \request("title_{$x}"),
-                    "answer" => \request("answer_{$x}"),
+                    'question' => \request("title_{$x}"),
+                    "answer" => \request(\request("answer_{$x}")),
+                    'test' => json_encode($test),
                     'test_id' => $id,
                 ]);
             }
@@ -102,9 +142,23 @@ class AdminController extends Controller
                 ]);
                 $up = Exam::find(Exam::where('test_id', $id)->get('id')[$x - 1]->id);
                 redirect()->back()->withErrors($validate);
+                $test = [];
+                if (strlen(\request("test_{$x}_1")) != 0 || \request("answer_{$x}") =="test_{$x}_1") {
+                    array_push($test, \request("test_{$x}_1"));
+                };
+                if (strlen(\request("test_{$x}_2")) != 0 || \request("answer_{$x}") =="test_{$x}_2") {
+                    array_push($test, \request("test_{$x}_2"));
+                };
+                if (strlen(\request("test_{$x}_3")) != 0 || \request("answer_{$x}") =="test_{$x}_3") {
+                    array_push($test, \request("test_{$x}_3"));
+                };
+                if (strlen(\request("test_{$x}_4")) != 0 || \request("answer_{$x}") =="test_{$x}_4") {
+                    array_push($test, \request("test_{$x}_4"));
+                };
                 $up->update([
-                    'qu' => \request("title_{$x}"),
-                    "answer" => \request("answer_{$x}"),
+                    'question' => \request("title_{$x}"),
+                    "answer" => \request(\request("answer_{$x}")),
+                    'test' => json_encode($test),
                     'test_id' => $id,
                 ]);
             }
